requirements.txt
blinker==1.9.0
boto3==1.42.17
botocore==1.42.17
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cycler>=0.12.1
docx2txt>=0.8
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib>=1.3.2
kiwisolver>=1.4.5
MarkupSafe==3.0.3
matplotlib>=3.8.2
numpy>=1.26.2
pandas>=2.1.4
pdfminer.six>=20231228
Pillow>=10.1.0
pycryptodome>=3.19.1
pyEDFlib>=0.1.36
pyparsing>=3.1.1
python-dateutil>=2.8.2
python-dotenv==1.2.1
pytz>=2023.3.post1
requests==2.32.5
s3transfer==0.16.0
scikit-learn==1.7.2
scipy>=1.11.4
seaborn>=0.13.0
six>=1.16.0
urllib3==2.6.2
Werkzeug==3.1.4
